Checksum checks for reports

+++ elements/IGMPRouter.cc
@@ -108,23 +108,30 @@ void IGMPRouter::received_igmp_query(int port, Packet* p) {
 
 void IGMPRouter::received_igmp_report(int port, Packet* p) {
 
-    // TODO try catch zoals client
-    IPAddress source = p->ip_header()->ip_src;
+    try {
+        if (not checkQuery(p)) {
+            p->kill();
+            return;
+        }
+        IPAddress source = p->ip_header()->ip_src;
+        // parse report
+        MembershipReport* report = (MembershipReport*) (p->data() + p->ip_header_length());
+        // get group records
+        uint16_t n = ntohs(report->n_group_records);
+        auto records = report->get_group_records(p, n);
 
-    // make report
-    MembershipReport* report = (MembershipReport*) (p->data() + p->ip_header_length());
-    // get group records
-    uint16_t n = ntohs(report->n_group_records);
-    auto records = report->get_group_records(p, n);
+        for (int i = 0; i < n; i++) {
+            // zoek in table naar de group state
+            click_chatter(" - A group record: %s on port %d", records[i]->multicast_address.unparse().c_str(), port);
+            this->state->RouterState::find_insert_group_state(port, source, records[i]->multicast_address);
+        }
+        p->kill();
 
-    for (int i = 0; i < n; i++) {
-        // zoek in table naar de group state
-        click_chatter(" - A group record: %s on port %d", records[i]->multicast_address.unparse().c_str(), port);
-        this->state->RouterState::find_insert_group_state(port, source, records[i]->multicast_address);
     }
+    catch (...) {
+        p->kill();
 
-    click_chatter("Router:\tHandled Report. Size of records: %d", records.size());
-    p->kill();
+    }
 }
 
 void IGMPRouter::add_handlers() {
@@ -168,6 +175,21 @@ void IGMPRouter::send_group_specific_query(Timer* timer, void* ptr) {
     query.set_max_response_code(router->last_member_query_interval);
     query.group_address = router->_group_address;
     router->_group_address = IPAddress("0.0.0.0");
+
+}
+
+bool IGMPRouter::checkQuery(Packet *p) {
+
+    // TODO IP header  checken
+    // TODO UDP header checken
+
+    MembershipReport* report = (MembershipReport*) (p->data() + p->ip_header_length());
+
+    //if (p->length() - p->ip_header_length() > sizeof(query)) { return false; }
+
+    // One's complement of total payload is done by adding the bitwise operator & 0xFFFF
+    // must be equal to 0
+    return (click_in_cksum((unsigned char*) report, p->length() - p->ip_header_length()) & 0xFFFF) == 0;
 
 }
